게시물 삭ì 제 이미 ì삭제 확인

+++ src/main/java/com/kny/exam/jpaBoard/acticle/controller/ArticleController.java
@@ -52,6 +52,10 @@ public class ArticleController {
     @RequestMapping("doDelete")
     @ResponseBody
     public String doDelete(long id){
+
+        if( articleRepository.existsById(id) == false) {
+            return "%d번 게시물은 이미 삭제 되었습니다.".formatted(id);
+        }
         articleRepository.deleteById(id);
         return "%d번 게시물이 삭제 되었습니다.".formatted(id);
     }
